added ui updates for when generating/saving/deleting images

+++ src/components/Entity.jsx
@@ -13,6 +13,7 @@ import { useEffect } from "react";
 import { availableVoices, delimiters } from "../utils/constants";
 import { generateImage, generateVoice } from "../utils/generation";
 import { deleteFile, uploadFile } from "../utils/storageUtils";
+import { useState } from "react";
 
 const Entity = ({
   index,
@@ -26,7 +27,10 @@ const Entity = ({
     useContext(ApplicationContext);
 
   let audioPlayer = null;
-  const [shouldDelete, setShouldDelete] = React.useState(false);
+  const [shouldDelete, setShouldDelete] = useState(false);
+  const [generatingImage, setGeneratingImage] = useState(false);
+  const [savingImage, setSavingImage] = useState(false);
+  const [cancelingImage, setCancelingImage] = useState(false);
 
   const updateEntity = (entities, field, data, index) => {
     if (field) {
@@ -50,11 +54,7 @@ const Entity = ({
       //     "#" +
       //     data;
       // }
-
-      if (field === "image") {
-        newData["imageCid"] = "";
-      }
-
+      
       if (!field) {
         newData = data;
       }
@@ -176,21 +176,33 @@ const Entity = ({
   };
 
   const deleteImage = async () => {
-    if (data["image"]?.length === 59) {
+    console.log("deleting image");
+    setCancelingImage(true);
+    if (data["image"]?.length === 59 && !cancelingImage) {
       await deleteFile(web3Storage, data["image"]);
     }
-    updateEntity(data, "image", "", index);
-    updateEntity(data, "imageCid", "", index);
+
+    let newData = { ...data };
+    newData["image"] = "";
+    newData["imageCid"] = "";
+    editEntityCallback(newData, index);
+    setCancelingImage(false);
+    console.log("deleted");
   };
 
   const saveImage = async () => {
-    if (!data["image"] || data["imageCid"]?.length === 59) {
-      console.log(data["imageCid"]?.length);
+    if (savingImage) {
       return;
     }
 
+    if (!data["image"] || data["imageCid"]?.length === 59) {
+      return;
+    }
+
+    setSavingImage(true);
     const cid = await uploadFile(web3Storage, data["image"]);
     updateEntity(data, "imageCid", cid, index);
+    setSavingImage(false);
   };
 
   return (
@@ -245,15 +257,21 @@ const Entity = ({
                 <div key={i}>
                   {data[field]?.length > 0 && (
                     <div>
-                      <button onClick={deleteImage}>Cancel</button>
-                      <button onClick={saveImage}>Save</button>
+                      <button onClick={deleteImage}>
+                        {cancelingImage ? "Canceling..." : "Cancel"}
+                      </button>
+                      <button onClick={saveImage}>
+                        {savingImage ? "Saving..." : "Save"}
+                      </button>
                     </div>
                   )}
                   <button
                     onClick={async () => {
-                      console.log(
-                        data["image"].length > 0 || data["imageCid"].length > 0
-                      );
+                      if (generatingImage) {
+                        return;
+                      }
+
+                      setGeneratingImage(true);
                       if (
                         data["image"].length > 0 ||
                         data["imageCid"].length > 0
@@ -269,9 +287,12 @@ const Entity = ({
                         ),
                         index
                       );
+                      setGeneratingImage(false);
                     }}
                   >
-                    {data[field]?.length > 0
+                    {generatingImage
+                      ? "Generating..."
+                      : data[field]?.length > 0
                       ? "Regenerate Image"
                       : "Generate Image"}
                   </button>
